remote_server/requirements.txt
astroid==1.6.1
beautifulsoup4==4.6.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
nltk==3.2.5
plumbum==1.6.6
pylint==1.8.2
python-crontab==2.2.8
python-dateutil==2.7.2
rpyc==3.4.4
six==1.11.0
tinydb==3.9.0
wrapt==1.10.11
